Added file support for extra.lua commands
Reply to a file (sticker, gif, document, audio, voice...) with an
optional text after the extra command and it will send that file to the
chat. Stickers can't have extra captions like the other files.
Example: /extra /cat text <reply a photo> and when you enter: /cat the
bot will send the photo with the caption